fix type in health meter

+++ lib/components/health_meter.dart
@@ -23,7 +23,7 @@ class HealthMeter extends StatelessWidget {
         children: [
           const Text("Health", textAlign: TextAlign.center),
           CircularMeter(
-            value: value,
+            value: value.toInt(),
             centerWidget: const Icon(
               Icons.favorite,
               color: Colors.red,
